refactor(player-inline): refactor player-inline module
Linked to #33
