+++ public/config/config.js
@@ -1,9 +1,25 @@
 import angular from 'angular';
+import AppLoadingBar from './loading.config';
+import uiBootstrap from 'angular-bootstrap';
+import ngFileUpload from 'ng-file-upload';
+import PlayerInlineModule from 'common/component/player-inline/player-inline';
+import 'angular-touch';
+import 'angular-animate';
+import 'angular-truncate';
+
+
+console.log(ngFileUpload);
 
 export default angular.module('ps.config', [
-    'ps.config.loading',
-    'ps.config.restangular',
-    'ps.config.ngstomp',
-    'ps.config.module',
-    'ps.partial'
+    AppLoadingBar.name,
+    'ngTouch',
+    'ngAnimate',
+    uiBootstrap,
+    'truncate',
+    ngFileUpload,
+    'ps.common.component.players-inline',
+    'ps.common.component.navbar',
+    'ps.common.component.authorize-notification',
+    'ps.common.component.device-detection',
+    'ps.common.component.updating'
 ]);
