+++ public/config/loading.config.js
@@ -1,3 +1,5 @@
-angular
+import 'angular-loading-bar';
+
+export default angular
     .module('ps.config.loading', [ 'angular-loading-bar' ])
     .config((cfpLoadingBarProvider) =>  { cfpLoadingBarProvider.includeSpinner = false; } );
