+++ public/config/truncate.js
@@ -1,0 +1,8 @@
+/**
+    * Created by kevin on 25/10/2015 for PodcastServer
+    */
+import 'angular-truncate';
+
+export default {
+    name : 'truncate'
+}
